Дуже багато всього змінено в модалках NEED HELP, ADD/EDIT BOARD, EDIT PROFILE, доданий один каркас на всіх, багато змін в стилях, багато змін в логіці роботи ADD/EDIT BOARD
