+++ src/components/ModalBoard/data.jsx
@@ -10,19 +10,19 @@ export const arrIcons = [
 ];
 
 export const arrBG = [
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-22_w6d1ix.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-21_lhyxtz.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-23_nvf3nr.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-19_jgsoar.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-18_lz8hnp.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-16_hbougf.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-15_p5qieq.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-13_kzosan.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-12_nehdkx.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-24_ob1z2m.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-25_cdzpq5.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060747/teamProject/bcg-laptop-20_je1cm3.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-17_ypzhw7.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-14_gvggrl.jpg',
-  'https://res.cloudinary.com/dpurp913m/image/upload/v1693060746/teamProject/bcg-laptop-11_rfi4bm.jpg',
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617335/bg-img/mobile/1/1-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617336/bg-img/mobile/1/2-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617337/bg-img/mobile/1/3-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617339/bg-img/mobile/1/4-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617340/bg-img/mobile/1/5-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617341/bg-img/mobile/1/6-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617342/bg-img/mobile/1/7-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617343/bg-img/mobile/1/8-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617344/bg-img/mobile/1/9-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617345/bg-img/mobile/1/10-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617346/bg-img/mobile/1/11-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617347/bg-img/mobile/1/12-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617347/bg-img/mobile/1/13-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617349/bg-img/mobile/1/14-bg-mobile-1x.jpg",
+  "https://res.cloudinary.com/teamprojectavatar/image/upload/v1693617350/bg-img/mobile/1/15-bg-mobile-1x.jpg"
 ];
